Continued refactoring of index.js
