+++ src/__tests__/Radi.test.js
@@ -1,0 +1,14 @@
+import Radi from '../Radi';
+
+describe('Radi.js', () => {
+  it('doesn\'t crash', () => {
+    new Radi({
+      state: {},
+      props: {},
+      actions: {},
+      $mixins: {},
+      $view: () => {},
+      name: 'foobar'
+    });
+  });
+});
